:ambulance: fixed return from endpoints to get data only to the user
close #22

+++ lib/comadrepay_web/controllers/user_controller.ex
@@ -26,18 +26,30 @@ defmodule ComadrepayWeb.UserController do
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Accounts.get_user!(id)
+    user_resource = Guardian.Plug.current_resource(conn)
 
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
-      render(conn, "show.json", user: user)
+    if user_resource.id == id do
+      user = Accounts.get_user!(id)
+
+      with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+        render(conn, "show.json", user: user)
+      end
+    else
+      {:error, :belong_user}
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
+    %{id: user_id} = Guardian.Plug.current_resource(conn)
 
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+    if user_id == id do
+      user = Accounts.get_user!(id)
+
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      {:error, :belong_user}
     end
   end
